+++ packages/sdks-tests/src/specs/columns-vertical-center-flex.ts
@@ -1,0 +1,118 @@
+export const COLUMNS_VERTICAL_CENTER_FLEX = {
+  ownerId: 'db60bf3db7fa4db7be81ef05b72bd720',
+  lastUpdateBy: null,
+  createdDate: 1735358202700,
+  id: '9f94156929a24d35ba1cb3ded92fd44c',
+  '@version': 4,
+  name: 'Vertical Centering Issue',
+  modelId: '144e96dac3294e7c8249a27fb11fdcb7',
+  published: 'draft',
+  meta: {
+    hasLinks: false,
+    kind: 'page',
+    lastPreviewUrl:
+      'http://localhost:4200/vertical-centering-issue?builder.space=db60bf3db7fa4db7be81ef05b72bd720&builder.user.permissions=read%2Ccreate%2Cpublish%2CeditDesigns%2CeditLayouts%2CeditLayers%2CeditContentPriority&builder.user.role.name=Designer&builder.user.role.id=creator&builder.cachebust=true&builder.preview=page&builder.noCache=true&builder.allowTextEdit=true&__builder_editing__=true&builder.overrides.page=9f94156929a24d35ba1cb3ded92fd44c&builder.overrides.9f94156929a24d35ba1cb3ded92fd44c=9f94156929a24d35ba1cb3ded92fd44c&builder.overrides.page:/vertical-centering-issue=9f94156929a24d35ba1cb3ded92fd44c',
+  },
+  priority: -273,
+  query: [
+    {
+      '@type': '@builder.io/core:Query',
+      property: 'urlPath',
+      operator: 'is',
+      value: '/vertical-centering-issue',
+    },
+  ],
+  data: {
+    title: 'Vertical Centering Issue',
+    themeId: false,
+    inputs: [],
+    blocks: [
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        id: 'builder-498c8104b0c34b66a9242eca717d63e5',
+        component: {
+          name: 'Columns',
+          options: {
+            columns: [
+              {
+                blocks: [
+                  {
+                    '@type': '@builder.io/sdk:Element',
+                    '@version': 2,
+                    id: 'builder-85fb71a32e68408dbaa2e86f8e3428ab',
+                    responsiveStyles: {
+                      large: {
+                        display: 'flex',
+                        flexDirection: 'column',
+                        position: 'relative',
+                        flexShrink: '0',
+                        boxSizing: 'border-box',
+                        marginTop: '20px',
+                        height: '200px',
+                      },
+                    },
+                  },
+                ],
+              },
+              {
+                blocks: [
+                  {
+                    '@type': '@builder.io/sdk:Element',
+                    '@version': 2,
+                    id: 'builder-17df9e074c3a4c00b7208a9163dcf535',
+                    component: {
+                      name: 'Text',
+                      options: {
+                        text: 'Enter some text...',
+                      },
+                    },
+                    responsiveStyles: {
+                      large: {
+                        display: 'flex',
+                        flexDirection: 'column',
+                        position: 'relative',
+                        flexShrink: '0',
+                        boxSizing: 'border-box',
+                        marginTop: 'auto',
+                        lineHeight: 'normal',
+                        height: 'auto',
+                        marginBottom: 'auto',
+                      },
+                    },
+                  },
+                ],
+              },
+            ],
+            space: 20,
+            stackColumnsAt: 'tablet',
+            reverseColumnsWhenStacked: false,
+          },
+        },
+        responsiveStyles: {
+          large: {
+            display: 'flex',
+            flexDirection: 'column',
+            position: 'relative',
+            flexShrink: '0',
+            boxSizing: 'border-box',
+            marginTop: '20px',
+            backgroundColor: 'rgba(120, 118, 118, 1)',
+          },
+        },
+      },
+    ],
+  },
+  metrics: {
+    clicks: 0,
+    impressions: 0,
+  },
+  variations: {},
+  lastUpdated: 1738142770171,
+  testRatio: 1,
+  screenshot:
+    'https://cdn.builder.io/api/v1/image/assets%2Fdb60bf3db7fa4db7be81ef05b72bd720%2F8a2a93f50b4d4e02996fbcc055209281',
+  createdBy: 'xStHbrbPN3W3WbSIHaJK1oGoFHG2',
+  lastUpdatedBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+  folders: [],
+};