+++ packages/sdks/src/scripts/init-editing.ts
@@ -26,14 +26,14 @@ export const registerInsertMenu = () => {
 };
 
 let isSetupForEditing = false;
-export const setupBrowserForEditing = (
-  options: {
-    enrich?: boolean;
-    includeRefs?: boolean;
-    locale?: string;
-    trustedHosts?: string[];
-  } = {}
-) => {
+export const setupBrowserForEditing = (options: {
+  modelName: string;
+  apiKey: string;
+  enrich?: boolean;
+  includeRefs?: boolean;
+  locale?: string;
+  trustedHosts?: string[];
+}) => {
   if (isSetupForEditing) {
     return;
   }
@@ -50,6 +50,10 @@ export const setupBrowserForEditing = (
           // scope our '+ add block' button styling
           supportsAddBlockScoping: true,
           supportsCustomBreakpoints: true,
+          modelName: options.modelName,
+          apiKey: options.apiKey,
+          supportsXSmallBreakpoint: TARGET === 'reactNative' ? false : true,
+          blockLevelPersonalization: true,
         },
       },
       '*'