+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @builder.io/react
 
+## 8.0.9
+
+### Patch Changes
+
+- 3c77cc7: chore: send `apiKey` to Visual Editor to improve editing experience.
+
+## 8.0.8
+
+### Patch Changes
+
+- abe5cba: Fix: hydration mismatch error and reactivity of Personalization Containers when `userAttributes` cookie value is updated.
+
 ## 8.0.7
 
 ### Patch Changes